requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
tqdm==4.66.5
scikit-learn==1.5.1
opencv-python==4.10.0.84
tensorflow==2.15.0
Pillow==10.4.0
